Revert "Add support of AMQP queues (#1063)"
This reverts commit efec9c04a3c4f881dc27e1d293d26705a3bc0980.
