+++ contribs/build.gradle
@@ -13,8 +13,6 @@ dependencies {
 
 	compile "io.nats:java-nats-streaming:${revNatsStreaming}"
 
-	compile "com.rabbitmq:amqp-client:${revAmqpClient}"
-
 	compileOnly "javax.ws.rs:jsr311-api:${revJsr311Api}"
 	compile "io.swagger:swagger-jaxrs:${revSwagger}"
 
@@ -22,5 +20,4 @@ dependencies {
 
 	testCompile "org.eclipse.jetty:jetty-server:${revJetteyServer}"
 	testCompile "org.eclipse.jetty:jetty-servlet:${revJettyServlet}"
-	testCompile "org.slf4j:slf4j-log4j12:${revSlf4jlog4j}"
 }
