+++ server/src/main/resources/server.properties
@@ -44,22 +44,3 @@ EC2_AVAILABILITY_ZONE=us-east-1c
 
 # disable archival service
 workflow.archive=false
-
-# OPTIONAL (contribs) - Settings for AMQP queues (RabbitMQ)
-# You must add module AMQModule to enable support of AMQP queues
-# Here are the settings with default values:
-# workflow.event.queues.amqp.hosts=localhost
-# workflow.event.queues.amqp.username=guest
-# workflow.event.queues.amqp.password=guest
-# workflow.event.queues.amqp.virtualHost=/
-# workflow.event.queues.amqp.port=5672
-# workflow.event.queues.amqp.useNio=false
-# workflow.event.queues.amqp.batchSize=1
-# workflow.event.queues.amqp.pollTimeInMs=100
-# Use durable queue ?
-# workflow.event.queues.amqp.durable=true
-# Use exclusive queue ?
-# workflow.event.queues.amqp.exclusive=false
-# Enable support of priorities on queue. Set the max priority on message.
-# Setting is ignored if the value is lower or equals to 0
-# workflow.event.queues.amqp.maxPriority=-1
